[feat] AdminMainView에서 MensStatusView 분리
